Initial support for DirectX 9 (#3)
* Initial attempt at upgrading to DirectX 9
* Renderer `ww3d2` now compiles

**Notes:**
* Untested as none of the game apps which use DirectX compile yet
* Similar to C++ language upgrade, it's likely a lot of the changes here
aren't correct but I'll fix the issue as/when they appear

+++ Code/ww3d2/render2dsentence.cpp
@@ -1524,8 +1524,8 @@ FontCharsClass::Grow_Unicode_Array (WCHAR ch)
 		return ;
 	}
 
-	uint16 first_index	= min( FirstUnicodeChar, ch );
-	uint16 last_index		= max( LastUnicodeChar, ch );
+	uint16 first_index	= min( static_cast<uint16_t>(FirstUnicodeChar), static_cast<uint16_t>(ch) );
+	uint16 last_index		= max( static_cast<uint16_t>(LastUnicodeChar), static_cast<uint16_t>(ch) );
 	uint16 count			= (last_index - first_index) + 1;
 
 	//
